outsourced api from main to api.py
